+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+ - new action for clustering nuvlaboxes
+ - new action for cleaning up leftover NuvlaBox clusters
+
 ## [2.14.2] - 2021-04-15
 
 ### Changed